Rewrite alternative filter operator support
- simplify scalar filtering support
- create FilterSpec to allow users to define alternative filters
- remove experimental filter operator functions
- create types to make src/filters.ts logic easier to reason about

Writing a typesafe interface for alternative filters is not straight
forward. This is due to the fact that a filter's correctness depends on
the state of the Postgres database. That necessitates checking to be
done at runtime. An additional complication is that a filter operator
may be valid for a wide range of types.

Because of the above contraints, we enforce an interface that is either
a scalar, array of scalars or a FilterSpec. The FilterSpec is what
allows alternative operators. The FilterSpec interface ensures that the
operator is valid, but does not otherwise make any attempt to convert or
modify the value. Thus, it is on the user to make sure the operator they
choose is valid for both the field they are filtering on and the value
they are filtering with.